fix: backend route permissions

+++ api/firstpassapp/views.py
@@ -96,9 +96,12 @@ def create_vault(request):
     data = json.loads(request.body)
     name = data.get('name')
     path = data.get('path')
-    userID = data.get('userID')
+    
+    user = request.user
+    if not user.is_authenticated:
+        return JsonResponse(data={'error': 'User not authenticated'}, status=401)
+    
     vault = Vault.objects.create(name=name, image_path=path)
-    user = User.objects.get(id=userID)
     account = Account.objects.get(user=user)
     # vault = Vault.objects.get(name=name, image_path=path)
     AccountVaultAccess.objects.create(
@@ -124,6 +127,11 @@ def get_vaults_for_user(request, user_id):
     account = Account.objects.get(user=user)
     account_vaults = AccountVaultAccess.objects.filter(account=account)
     vaults = []
+    
+    user = request.user
+    if not user.is_authenticated:
+        return JsonResponse(data={'error': 'User not authenticated'}, status=401)
+    
     for account_vault in account_vaults:
         vaults.append(account_vault.vault)
     jsonVaults = []
@@ -139,6 +147,11 @@ def get_invitations_for_user(request, user_id):
     account = Account.objects.get(user=user)
     invitations = Invitation.objects.filter(account=account)
     jsonInvitations = []
+    
+    user = request.user
+    if not user.is_authenticated:
+        return JsonResponse(data={'error': 'User not authenticated'}, status=401)
+    
     for invitation in invitations:
         jsonInvitations.append({'id': invitation.id, 'account': invitation.account.id,
                                'vault': invitation.vault.id, 'access_level': invitation.access_level})
@@ -151,6 +164,13 @@ def accept_invitation(request, invitation_id):
     account = invitation.account
     vault = invitation.vault
     access_level = invitation.access_level
+    
+    user = request.user
+    if not user.is_authenticated:
+        return JsonResponse(data={'error': 'User not authenticated'}, status=401)
+    if user.id != account.user.id:
+        return JsonResponse(data={'error': 'User not authorized'}, status=403)
+    
     AccountVaultAccess.objects.create(
         account=account, vault=vault, access_level=access_level)
     invitation.delete()
@@ -160,6 +180,14 @@ def accept_invitation(request, invitation_id):
 @require_POST
 def decline_invitation(request, invitation_id):
     invitation = Invitation.objects.get(id=invitation_id)
+    
+    user = request.user
+    if not user.is_authenticated:
+        return JsonResponse(data={'error': 'User not authenticated'}, status=401)
+    
+    if user.id != invitation.account.user.id:
+        return JsonResponse(data={'error': 'User not authorized'}, status=403)
+    
     invitation.delete()
     return JsonResponse(data={'invitation': invitation.id}, status=200)
 
@@ -195,6 +223,11 @@ def route_invitations(request):
 def get_invitations(request):
     invitations = Invitation.objects.all()
     jsonInvitations = []
+    
+    user = request.user
+    if not user.is_authenticated:
+        return JsonResponse(data={'error': 'User not authenticated'}, status=401)
+    
     for invitation in invitations:
         jsonInvitations.append({'id': invitation.id, 'account': invitation.account.id,
                                'vault': invitation.vault.id, 'access_level': invitation.access_level})
@@ -204,13 +237,22 @@ def get_invitations(request):
 @require_POST
 def send_invitation(request):
     data = json.loads(request.body)
-    accountID = data.get('accountId')
-    vaultID = data.get('vaultId')
-    accessLevel = data.get('accessLevel')
-    account = Account.objects.get(id=accountID)
-    vault = Vault.objects.get(id=vaultID)
+    account_id = data.get('accountId')
+    vault_id = data.get('vaultId')
+    access_level = data.get('accessLevel')
+    account = Account.objects.get(id=account_id)
+    vault = Vault.objects.get(id=vault_id)
+    
+    user = request.user
+    if not user.is_authenticated:
+        return JsonResponse(data={'error': 'User not authenticated'}, status=401)    
+    
+    access_level_invitor = AccountVaultAccess.objects.get(account=user.account, vault=vault).access_level
+    if access_level_invitor != 'O':
+        return JsonResponse(data={'erreor': 'User not authorized'}, status=403)
+    
     invitation = Invitation.objects.create(
-        account=account, vault=vault, access_level=accessLevel)
+        account=account, vault=vault, access_level=access_level)
     jsonInvitation = {'id': invitation.id, 'account': invitation.account.id,
                       'vault': invitation.vault.id, 'access_level': invitation.access_level}
     return JsonResponse(data={'invitation': jsonInvitation}, status=200)
@@ -232,6 +274,11 @@ def route_vaults(request, vault_id):
 @require_GET
 def get_vault_by_id(request, vault_id):
     vault = Vault.objects.get(id=vault_id)
+    
+    user = request.user
+    if not user.is_authenticated:
+        return JsonResponse(data={'error': 'User not authenticated'}, status=401)
+    
     jsonVault = {'id': vault.id, 'name': vault.name,
                  'image_path': vault.image_path}
     return JsonResponse(data={'vault': jsonVault}, status=200)
@@ -245,6 +292,15 @@ def update_vault_by_id(request, vault_id):
     vault = Vault.objects.get(id=vault_id)
     vault.name = name
     vault.image_path = image_path
+    
+    user = request.user
+    if not user.is_authenticated :
+        return JsonResponse(data={'error': 'User not authenticated'}, status=401)
+
+    access_level = AccountVaultAccess.objects.get(account=user.account, vault=vault).access_level
+    if user.account not in vault.users.all() or access_level != 'O':
+        return JsonResponse(data={'error': 'User does not have access to this vault'}, status=403)
+    
     vault.save()
     jsonVault = {'id': vault.id, 'name': vault.name,
                  'image_path': vault.image_path}
@@ -258,9 +314,8 @@ def delete_vault_by_id(request, vault_id):
     if not user.is_authenticated:
         return JsonResponse(data={'error': 'User not authenticated'}, status=401)
 
-    access_level = AccountVaultAccess.objects.get(
-        account=user.account, vault=vault).access_level
-    if not (user in vault.users.all() or access_level != 'owner'):
+    access_level = AccountVaultAccess.objects.get(account=user.account, vault=vault).access_level
+    if user.account not in vault.users.all() or access_level != 'O' :
         return JsonResponse(data={'error': 'User does not have access to this vault'}, status=403)
 
     vault.delete()
@@ -280,6 +335,14 @@ def get_pairs(request, vault_id):
     vault = Vault.objects.get(id=vault_id)
     pairs = Pair.objects.filter(vault=vault)
     jsonPairs = []
+    
+    user = request.user
+    if not user.is_authenticated:
+        return JsonResponse(data={'error': 'User not authenticated'}, status=401)
+    
+    if user.account not in vault.users.all():
+        return JsonResponse(data={'error': 'User does not have access to this vault'}, status=403)
+    
     for pair in pairs:
         jsonPairs.append({'id': pair.id, 'application': pair.application,
                           'username': pair.username, 'password': pair.password})
@@ -296,10 +359,9 @@ def add_pair(request, vault_id):
     user = request.user
     if not user.is_authenticated:
         return JsonResponse(data={'error': 'User not authenticated'}, status=401)
-
-    access_level = AccountVaultAccess.objects.get(
-        account=user.account, vault=vault).access_level
-    if not (user in vault.users.all() or (access_level != 'owner' and access_level != 'write')):
+    
+    access_level = AccountVaultAccess.objects.get(account=user.account, vault=vault).access_level
+    if user.account not in vault.users.all() or (access_level != 'O' and access_level != 'W'):
         return JsonResponse(data={'error': 'User does not have access to this vault'}, status=403)
 
     pair = Pair.objects.create(
@@ -323,9 +385,9 @@ def update_pair(request, pair_id):
     user = request.user
     if not user.is_authenticated:
         return JsonResponse(data={'error': 'User not authenticated'}, status=401)
-    access_level = AccountVaultAccess.objects.get(
-        account=user.account, vault=vault).access_level
-    if not (user in vault.users.all() or (access_level != 'owner' and access_level != 'write')):
+
+    access_level = AccountVaultAccess.objects.get(account=user.account, vault=vault).access_level
+    if user.account not in vault.users.all() or (access_level != 'O' and access_level != 'W'):
         return JsonResponse(data={'error': 'User does not have access to this vault'}, status=403)
 
     pair.application = application
@@ -349,9 +411,9 @@ def delete_pair(request, pair_id):
     user = request.user
     if not user.is_authenticated:
         return JsonResponse(data={'error': 'User not authenticated'}, status=401)
-    access_level = AccountVaultAccess.objects.get(
-        account=user.account, vault=vault).access_level
-    if not (user in vault.users.all() or (access_level != 'owner' and access_level != 'write')):
+
+    access_level = AccountVaultAccess.objects.get(account=user.account, vault=vault).access_level
+    if user.account not in vault.users.all() or (access_level != 'O' and access_level != 'W'):
         return JsonResponse(data={'error': 'User does not have access to this vault'}, status=403)
 
     pair = Pair.objects.get(id=pair_id)
@@ -367,6 +429,10 @@ def update_user(request, user_id):
     new_password = data.get('new_password')
     confirm_password = data.get('confirm_password')
     user = User.objects.get(id=user_id)
+    
+    user = request.user
+    if not user.is_authenticated:
+        return JsonResponse(data={'error': 'User not authenticated'}, status=401)
     if user.check_password(old_password):
         if new_password == confirm_password:
             user.username = username
